requirements.txt
absl-py==0.12.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2023.7.22
charset-normalizer==2.0.12
clang==5.0
colorama==0.4.5
cycler==0.11.0
dataclasses==0.8
dill==0.3.4
flatbuffers==1.12
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.3
grpcio==1.48.2
h5py==3.1.0
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
keras==2.6.0
keras-cv-attention-models==1.3.18
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
pandas==0.25.3
Pillow==8.4.0
promise==2.3
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2023.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
six==1.15.0
sklearn==0.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.0
tensorflow-addons==0.14.0
tensorflow-datasets==4.5.2
tensorflow-estimator==2.8.0
tensorflow-metadata==1.2.0
termcolor==1.1.0
threadpoolctl==3.1.0
tqdm==4.64.1
typeguard==2.13.3
typing-extensions==3.7.4.3
urllib3==1.26.16
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.6.0
